Start refactor of playground tests
