+++ model/project.go
@@ -53,16 +53,17 @@ func (p *Project) IsOwnedBy(userID uuid.UUID) bool {
 // and marks it as mutable.
 func (p *Project) ExportPublicMutable() *Project {
 	return &Project{
-		ID:          p.ID,
-		Title:       p.Title,
-		Description: p.Description,
-		Readme:      p.Readme,
-		PublicID:    p.PublicID,
-		ParentID:    p.ParentID,
-		Persist:     p.Persist,
-		Seed:        p.Seed,
-		Version:     p.Version,
-		Mutable:     true,
+		ID:               p.ID,
+		Title:            p.Title,
+		Description:      p.Description,
+		Readme:           p.Readme,
+		PublicID:         p.PublicID,
+		ParentID:         p.ParentID,
+		Persist:          p.Persist,
+		Seed:             p.Seed,
+		NumberOfAccounts: p.NumberOfAccounts,
+		Version:          p.Version,
+		Mutable:          true,
 	}
 }
 
@@ -70,16 +71,17 @@ func (p *Project) ExportPublicMutable() *Project {
 // and marks it as immutable.
 func (p *Project) ExportPublicImmutable() *Project {
 	return &Project{
-		ID:          p.ID,
-		Title:       p.Title,
-		Description: p.Description,
-		Readme:      p.Readme,
-		PublicID:    p.PublicID,
-		ParentID:    p.ParentID,
-		Persist:     p.Persist,
-		Seed:        p.Seed,
-		Version:     p.Version,
-		Mutable:     false,
+		ID:               p.ID,
+		Title:            p.Title,
+		Description:      p.Description,
+		Readme:           p.Readme,
+		PublicID:         p.PublicID,
+		ParentID:         p.ParentID,
+		Persist:          p.Persist,
+		Seed:             p.Seed,
+		NumberOfAccounts: p.NumberOfAccounts,
+		Version:          p.Version,
+		Mutable:          false,
 	}
 }
