--- playground_test.go
@@ -1,2778 +1,0 @@
-/*
- * Flow Playground
- *
- * Copyright 2019 Dapper Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package playground_test
-
-import (
-	"github.com/Masterminds/semver"
-	"github.com/dapperlabs/flow-playground-api/blockchain"
-	"github.com/dapperlabs/flow-playground-api/middleware/errors"
-	"github.com/getsentry/sentry-go"
-	"github.com/go-chi/chi"
-	"github.com/kelseyhightower/envconfig"
-	"github.com/sirupsen/logrus"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"strconv"
-	"strings"
-	"testing"
-
-	playground "github.com/dapperlabs/flow-playground-api"
-	"github.com/dapperlabs/flow-playground-api/auth"
-	legacyauth "github.com/dapperlabs/flow-playground-api/auth/legacy"
-	"github.com/dapperlabs/flow-playground-api/client"
-	"github.com/dapperlabs/flow-playground-api/middleware/httpcontext"
-	"github.com/dapperlabs/flow-playground-api/model"
-	"github.com/dapperlabs/flow-playground-api/storage"
-)
-
-type Project struct {
-	ID               string
-	Title            string
-	Description      string
-	Readme           string
-	Seed             int
-	Persist          bool
-	Version          string
-	NumberOfAccounts int
-	/*
-		Accounts    []struct { // TODO: Remove accounts and refactor all the tests :P
-			ID        string
-			Address   string
-			DraftCode string
-		}
-	*/
-	//ContractTemplates    []ContractTemplate
-	TransactionTemplates []TransactionTemplate
-	//ScriptTemplates      []ScriptTemplates
-	Secret string
-}
-
-const MutationCreateProject = `
-mutation($title: String!, $description: String!, $readme: String!, $seed: Int!, $numberOfAccounts: Int!, $transactionTemplates: [NewProjectTransactionTemplate!]) {
-  createProject(input: { title: $title, description: $description, readme: $readme, seed: $seed, numberOfAccounts: $numberOfAccounts, transactionTemplates: $transactionTemplates }) {
-    id
-    title
-		description
-		readme
-    seed
-    persist
-    version
-	numberOfAccounts
-    transactionTemplates {
-      id
-      title
-      script
-      index
-    }
-  }
-}
-`
-
-type CreateProjectResponse struct {
-	CreateProject Project
-}
-
-const QueryGetProject = `
-query($projectId: UUID!) {
-  project(id: $projectId) {
-    id
-    accounts {
-      id
-      address
-      draftCode
-      deployedCode
-    }
-  }
-}
-`
-
-type GetProjectResponse struct {
-	Project Project
-}
-
-const MutationUpdateProjectPersist = `
-mutation($projectId: UUID!, $title: String!, $description: String!, $readme: String!, $persist: Boolean!) {
-  updateProject(input: { id: $projectId, title: $title, description: $description, readme: $readme, persist: $persist }) {
-    id
-		title
-		description
-		readme
-    persist
-  }
-}
-`
-
-type UpdateProjectResponse struct {
-	UpdateProject struct {
-		ID          string
-		Title       string
-		Description string
-		Readme      string
-		Persist     bool
-	}
-}
-
-const QueryGetProjectTransactionTemplates = `
-query($projectId: UUID!) {
-  project(id: $projectId) {
-    id
-    transactionTemplates {
-      id
-      script
-      index
-    }
-  }
-}
-`
-
-type GetProjectTransactionTemplatesResponse struct {
-	Project struct {
-		ID                   string
-		TransactionTemplates []struct {
-			ID     string
-			Script string
-			Index  int
-		}
-	}
-}
-
-const QueryGetProjectScriptTemplates = `
-query($projectId: UUID!) {
-  project(id: $projectId) {
-    id
-    scriptTemplates {
-      id
-      script
-      index
-    }
-  }
-}
-`
-
-type GetProjectScriptTemplatesResponse struct {
-	Project struct {
-		ID              string
-		ScriptTemplates []struct {
-			ID     string
-			Script string
-			Index  int
-		}
-	}
-}
-
-type File struct {
-	ID     string
-	Title  string
-	Script string
-	Type   int
-	Index  int
-}
-
-type ContractTemplate = File
-type TransactionTemplate = File
-type ScriptTemplate = File
-
-const MutationCreateTransactionTemplate = `
-mutation($projectId: UUID!, $title: String!, $script: String!) {
-  createTransactionTemplate(input: { projectId: $projectId, title: $title, script: $script }) {
-    id
-    title
-    script
-    index
-  }
-}
-`
-
-type CreateTransactionTemplateResponse struct {
-	CreateTransactionTemplate TransactionTemplate
-}
-
-const QueryGetTransactionTemplate = `
-query($templateId: UUID!, $projectId: UUID!) {
-  transactionTemplate(id: $templateId, projectId: $projectId) {
-    id
-    script
-    index
-  }
-}
-`
-
-type GetTransactionTemplateResponse struct {
-	TransactionTemplate struct {
-		ID     string
-		Script string
-		Index  int
-	}
-}
-
-const MutationUpdateTransactionTemplateScript = `
-mutation($templateId: UUID!, $projectId: UUID!, $script: String!) {
-  updateTransactionTemplate(input: { id: $templateId, projectId: $projectId, script: $script }) {
-    id
-    script
-    index
-  }
-}
-`
-
-const MutationUpdateTransactionTemplateIndex = `
-mutation($templateId: UUID!, $projectId: UUID!, $index: Int!) {
-  updateTransactionTemplate(input: { id: $templateId, projectId: $projectId, index: $index }) {
-    id
-    script
-    index
-  }
-}
-`
-
-type UpdateTransactionTemplateResponse struct {
-	UpdateTransactionTemplate struct {
-		ID     string
-		Script string
-		Index  int
-	}
-}
-
-const MutationDeleteTransactionTemplate = `
-mutation($templateId: UUID!, $projectId: UUID!) {
-  deleteTransactionTemplate(id: $templateId, projectId: $projectId)
-}
-`
-
-type DeleteTransactionTemplateResponse struct {
-	DeleteTransactionTemplate string
-}
-
-const MutationCreateTransactionExecution = `
-mutation($projectId: UUID!, $script: String!, $signers: [Address!], $arguments: [String!]) {
-  createTransactionExecution(input: {
-    projectId: $projectId,
-    script: $script,
-    arguments: $arguments,
-    signers: $signers
-  }) {
-    id
-    script
-    errors {
-      message
-      startPosition { offset line column }
-      endPosition { offset line column }
-    }
-    logs
-    events {
-      type
-      values
-    }
-  }
-}
-`
-
-type CreateTransactionExecutionResponse struct {
-	CreateTransactionExecution struct {
-		ID     string
-		Script string
-		Errors []model.ProgramError
-		Logs   []string
-		Events []struct {
-			Type   string
-			Values []string
-		}
-	}
-}
-
-const MutationCreateScriptExecution = `
-mutation CreateScriptExecution($projectId: UUID!, $script: String!, $arguments: [String!]) {
-  createScriptExecution(input: {
-    projectId: $projectId,
-    script: $script,
-    arguments: $arguments
-  }) {
-    id
-    script
-    errors {
-      message
-      startPosition { offset line column }
-      endPosition { offset line column }
-    }
-    logs
-    value
-  }
-}
-`
-
-type CreateScriptExecutionResponse struct {
-	CreateScriptExecution struct {
-		ID     string
-		Script string
-		Errors []model.ProgramError
-		Logs   []string
-		Value  string
-	}
-}
-
-const MutationCreateContractTemplate = `
-mutation($projectId: UUID!, $title: String!, $script: String!) {
-  createContractTemplate(input: { projectId: $projectId, title: $title, script: $script }) {
-    id
-    title
-    script
-    index
-  }
-}
-`
-
-const MutationCreateContractDeployment = `
-mutation($projectId: UUID!, $script: String!, $address: Address!) {
-  createContractDeployment(input: { projectId: $projectId, script: $script, address: $address }) {
-    id
-    script
-    address
-  }
-}
-`
-
-const MutationCreateScriptTemplate = `
-mutation($projectId: UUID!, $title: String!, $script: String!) {
-  createScriptTemplate(input: { projectId: $projectId, title: $title, script: $script }) {
-    id
-    title
-    script
-    index
-  }
-}
-`
-
-type CreateContractDeploymentResponse struct {
-	CreateContractDeployment struct {
-		ID      string
-		Script  string
-		Address model.Address
-		Errors  []model.ProgramError
-		Events  []struct {
-			Type   string
-			Values []string
-		}
-		Logs []string
-	}
-}
-
-/*
-type ScriptTemplate struct {
-	ID     string
-	Title  string
-	Script string
-	Index  int
-}
-*/
-
-type CreateScriptTemplateResponse struct {
-	CreateScriptTemplate ScriptTemplate
-}
-
-const QueryGetScriptTemplate = `
-query($templateId: UUID!, $projectId: UUID!) {
-  scriptTemplate(id: $templateId, projectId: $projectId) {
-    id
-    script
-  }
-}
-`
-
-type GetScriptTemplateResponse struct {
-	ScriptTemplate ScriptTemplate
-}
-
-const MutationUpdateScriptTemplateScript = `
-mutation($templateId: UUID!, $projectId: UUID!, $script: String!) {
-  updateScriptTemplate(input: { id: $templateId, projectId: $projectId, script: $script }) {
-    id
-    script
-    index
-  }
-}
-`
-
-const MutationUpdateScriptTemplateIndex = `
-mutation($templateId: UUID!, $projectId: UUID!, $index: Int!) {
-  updateScriptTemplate(input: { id: $templateId, projectId: $projectId, index: $index }) {
-    id
-    script
-    index
-  }
-}
-`
-
-type UpdateScriptTemplateResponse struct {
-	UpdateScriptTemplate struct {
-		ID     string
-		Script string
-		Index  int
-	}
-}
-
-const MutationDeleteScriptTemplate = `
-mutation($templateId: UUID!, $projectId: UUID!) {
-  deleteScriptTemplate(id: $templateId, projectId: $projectId)
-}
-`
-
-type DeleteScriptTemplateResponse struct {
-	DeleteScriptTemplate string
-}
-
-const initAccounts = 5
-
-func TestReplicas(t *testing.T) {
-	// Each replica is a different client calling the API, but also an instance of the resolver
-	const numReplicas = 5
-
-	// Create replicas
-	var replicas []*Client
-	for i := 0; i < numReplicas; i++ {
-		replicas = append(replicas, newClient())
-	}
-
-	replicaIdx := 0 // current replica
-	// loadBalancer cycles through replicas
-	var loadBalancer = func() *Client {
-		replicaIdx = (replicaIdx + 1) % len(replicas)
-		return replicas[replicaIdx]
-	}
-
-	// Create project for all replica tests
-	c := loadBalancer()
-	project := createProject(t, c)
-	cookie := c.SessionCookie() // Use one session cookie for everything currently
-
-	t.Run("Execute transactions on multiple replicas", func(t *testing.T) {
-		for i := 0; i < 10; i++ {
-			const script = "transaction { execute { log(\"Hello, World!\") } }"
-			var resp CreateTransactionExecutionResponse
-			err := loadBalancer().Post(
-				MutationCreateTransactionExecution,
-				&resp,
-				client.Var("projectId", project.ID),
-				client.Var("script", script),
-				client.AddCookie(cookie),
-			)
-			require.NoError(t, err)
-			assert.Empty(t, resp.CreateTransactionExecution.Errors)
-			assert.Contains(t, resp.CreateTransactionExecution.Logs, "\"Hello, World!\"")
-			assert.Equal(t, script, resp.CreateTransactionExecution.Script)
-		}
-	})
-
-	t.Run("Re-deploy contracts on multiple replicas to initial accounts", func(t *testing.T) {
-		var contract = "pub contract Foo {}"
-
-		for i := 0; i < 10; i++ {
-			accountIdx := i % project.NumberOfAccounts
-
-			var deployResp CreateContractDeploymentResponse
-			err := loadBalancer().Post(
-				MutationCreateContractDeployment,
-				&deployResp,
-				client.Var("projectId", project.ID),
-				client.Var("address", model.NewAddressFromString("0x0"+strconv.Itoa(accountIdx))),
-				client.Var("code", contract),
-				client.AddCookie(cookie),
-			)
-			require.NoError(t, err)
-
-			/*
-				var updateResp UpdateAccountResponse
-				err := loadBalancer().Post(
-					MutationUpdateAccountDeployedCode,
-					&updateResp,
-					client.Var("projectId", project.ID),
-					client.Var("accountId", account.ID),
-					client.Var("code", contract),
-					client.AddCookie(cookie),
-				)
-				require.NoError(t, err)
-				assert.Equal(t, contract, updateResp.UpdateAccount.DeployedCode)
-				assert.Equal(t, updateResp.UpdateAccount.DeployedCode, contract)
-			*/
-		}
-	})
-}
-
-/*
-func TestProjects(t *testing.T) {
-	t.Parallel()
-
-	t.Run("Create empty project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp CreateProjectResponse
-
-		err := c.Post(
-			MutationCreateProject,
-			&resp,
-			client.Var("title", "foo"),
-			client.Var("description", "bar"),
-			client.Var("readme", "bah"),
-			client.Var("seed", 42),
-		)
-		require.NoError(t, err)
-
-		assert.NotEmpty(t, resp.CreateProject.ID)
-		assert.Equal(t, 42, resp.CreateProject.Seed)
-		assert.Equal(t, version.String(), resp.CreateProject.Version)
-
-		// project should be created with 4 default accounts
-		assert.Len(t, resp.CreateProject.Accounts, initAccounts)
-
-		// project should not be persisted
-		assert.False(t, resp.CreateProject.Persist)
-	})
-
-	t.Run("Create project with 2 accounts", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp CreateProjectResponse
-
-		accounts := []string{
-			"pub contract Foo {}",
-			"pub contract Bar {}",
-		}
-
-		err := c.Post(
-			MutationCreateProject,
-			&resp,
-			client.Var("title", "foo"),
-			client.Var("description", "desc"),
-			client.Var("readme", "rtfm"),
-			client.Var("seed", 42),
-			client.Var("accounts", accounts),
-		)
-		require.NoError(t, err)
-
-		// project should still be created with 4 default accounts
-		assert.Len(t, resp.CreateProject.Accounts, initAccounts)
-
-		assert.Equal(t, "0000000000000001", resp.CreateProject.Accounts[0].Address)
-		assert.Equal(t, "0000000000000002", resp.CreateProject.Accounts[1].Address)
-		assert.Equal(t, "0000000000000003", resp.CreateProject.Accounts[2].Address)
-		assert.Equal(t, "0000000000000004", resp.CreateProject.Accounts[3].Address)
-		assert.Equal(t, accounts[0], resp.CreateProject.Accounts[0].DraftCode)
-		assert.Equal(t, accounts[1], resp.CreateProject.Accounts[1].DraftCode)
-		assert.Equal(t, "", resp.CreateProject.Accounts[2].DraftCode)
-	})
-
-	t.Run("Create project with 4 accounts", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp CreateProjectResponse
-
-		accounts := []string{
-			"pub contract Foo {}",
-			"pub contract Bar {}",
-			"pub contract Dog {}",
-			"pub contract Cat {}",
-		}
-
-		err := c.Post(
-			MutationCreateProject,
-			&resp,
-			client.Var("title", "foo"),
-			client.Var("seed", 42),
-			client.Var("description", "desc"),
-			client.Var("readme", "rtfm"),
-			client.Var("accounts", accounts),
-		)
-		require.NoError(t, err)
-
-		// project should still be created with 4 default accounts
-		assert.Len(t, resp.CreateProject.Accounts, initAccounts)
-
-		assert.Equal(t, accounts[0], resp.CreateProject.Accounts[0].DraftCode)
-		assert.Equal(t, accounts[1], resp.CreateProject.Accounts[1].DraftCode)
-		assert.Equal(t, accounts[2], resp.CreateProject.Accounts[2].DraftCode)
-	})
-
-	t.Run("Create project with transaction templates", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp CreateProjectResponse
-
-		templates := []struct {
-			Title  string `json:"title"`
-			Script string `json:"script"`
-		}{
-			{
-				"foo", "transaction { execute { log(\"foo\") } }",
-			},
-			{
-				"bar", "transaction { execute { log(\"bar\") } }",
-			},
-		}
-
-		err := c.Post(
-			MutationCreateProject,
-			&resp,
-			client.Var("title", "foo"),
-			client.Var("seed", 42),
-			client.Var("description", "desc"),
-			client.Var("readme", "rtfm"),
-			client.Var("transactionTemplates", templates),
-		)
-		require.NoError(t, err)
-
-		assert.Len(t, resp.CreateProject.TransactionTemplates, 2)
-		assert.Equal(t, templates[0].Title, resp.CreateProject.TransactionTemplates[0].Title)
-		assert.Equal(t, templates[0].Script, resp.CreateProject.TransactionTemplates[0].Script)
-		assert.Equal(t, templates[1].Title, resp.CreateProject.TransactionTemplates[1].Title)
-		assert.Equal(t, templates[1].Script, resp.CreateProject.TransactionTemplates[1].Script)
-	})
-
-	t.Run("Get project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp GetProjectResponse
-
-		err := c.Post(
-			QueryGetProject,
-			&resp,
-			client.Var("projectId", project.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, project.ID, resp.Project.ID)
-	})
-
-	t.Run("Get non-existent project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp CreateProjectResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			QueryGetProject,
-			&resp,
-			client.Var("projectId", badID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Persist project without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp UpdateProjectResponse
-
-		err := c.Post(
-			MutationUpdateProjectPersist,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("title", project.Title),
-			client.Var("description", project.Description),
-			client.Var("readme", project.Readme),
-			client.Var("persist", true),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Persist project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp UpdateProjectResponse
-
-		err := c.Post(
-			MutationUpdateProjectPersist,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("title", project.Title),
-			client.Var("description", project.Description),
-			client.Var("readme", project.Readme),
-			client.Var("persist", true),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, project.ID, resp.UpdateProject.ID)
-		assert.Equal(t, project.Title, resp.UpdateProject.Title)
-		assert.Equal(t, project.Description, resp.UpdateProject.Description)
-		assert.Equal(t, project.Readme, resp.UpdateProject.Readme)
-		assert.True(t, resp.UpdateProject.Persist)
-	})
-}
-
-
-func TestTransactionTemplates(t *testing.T) {
-	t.Parallel()
-	t.Run("Create transaction template without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionTemplateResponse
-
-		err := c.Post(
-			MutationCreateTransactionTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "bar"),
-		)
-
-		assert.Error(t, err)
-		assert.Empty(t, resp.CreateTransactionTemplate.ID)
-	})
-
-	t.Run("Create transaction template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionTemplateResponse
-
-		err := c.Post(
-			MutationCreateTransactionTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "bar"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.NotEmpty(t, resp.CreateTransactionTemplate.ID)
-		assert.Equal(t, "foo", resp.CreateTransactionTemplate.Title)
-		assert.Equal(t, "bar", resp.CreateTransactionTemplate.Script)
-	})
-
-	t.Run("Get transaction template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateTransactionTemplateResponse
-
-		err := c.Post(
-			MutationCreateTransactionTemplate,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "bar"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		var respB GetTransactionTemplateResponse
-
-		err = c.Post(
-			QueryGetTransactionTemplate,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", respA.CreateTransactionTemplate.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, respA.CreateTransactionTemplate.ID, respB.TransactionTemplate.ID)
-		assert.Equal(t, respA.CreateTransactionTemplate.Script, respB.TransactionTemplate.Script)
-	})
-
-	t.Run("Get non-existent transaction template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp GetTransactionTemplateResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			QueryGetTransactionTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", badID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Update transaction template without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateTransactionTemplateResponse
-
-		err := c.Post(
-			MutationCreateTransactionTemplate,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "apple"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		templateID := respA.CreateTransactionTemplate.ID
-
-		var respB UpdateTransactionTemplateResponse
-
-		err = c.Post(
-			MutationUpdateTransactionTemplateScript,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.Var("script", "orange"),
-		)
-		assert.Error(t, err)
-	})
-
-	t.Run("Update transaction template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateTransactionTemplateResponse
-
-		err := c.Post(
-			MutationCreateTransactionTemplate,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "apple"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		templateID := respA.CreateTransactionTemplate.ID
-
-		var respB UpdateTransactionTemplateResponse
-
-		err = c.Post(
-			MutationUpdateTransactionTemplateScript,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.Var("script", "orange"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, respA.CreateTransactionTemplate.ID, respB.UpdateTransactionTemplate.ID)
-		assert.Equal(t, respA.CreateTransactionTemplate.Index, respB.UpdateTransactionTemplate.Index)
-		assert.Equal(t, "orange", respB.UpdateTransactionTemplate.Script)
-
-		var respC struct {
-			UpdateTransactionTemplate struct {
-				ID     string
-				Script string
-				Index  int
-			}
-		}
-
-		err = c.Post(
-			MutationUpdateTransactionTemplateIndex,
-			&respC,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.Var("index", 1),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, respA.CreateTransactionTemplate.ID, respC.UpdateTransactionTemplate.ID)
-		assert.Equal(t, 1, respC.UpdateTransactionTemplate.Index)
-		assert.Equal(t, respB.UpdateTransactionTemplate.Script, respC.UpdateTransactionTemplate.Script)
-	})
-
-	t.Run("Update non-existent transaction template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp UpdateTransactionTemplateResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			MutationUpdateTransactionTemplateScript,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", badID),
-			client.Var("script", "bar"),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Get transaction templates for project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		templateA := createTransactionTemplate(t, c, project)
-		templateB := createTransactionTemplate(t, c, project)
-		templateC := createTransactionTemplate(t, c, project)
-
-		var resp GetProjectTransactionTemplatesResponse
-
-		err := c.Post(
-			QueryGetProjectTransactionTemplates,
-			&resp,
-			client.Var("projectId", project.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Len(t, resp.Project.TransactionTemplates, 3)
-		assert.Equal(t, templateA.ID, resp.Project.TransactionTemplates[0].ID)
-		assert.Equal(t, templateB.ID, resp.Project.TransactionTemplates[1].ID)
-		assert.Equal(t, templateC.ID, resp.Project.TransactionTemplates[2].ID)
-
-		assert.Equal(t, 0, resp.Project.TransactionTemplates[0].Index)
-		assert.Equal(t, 1, resp.Project.TransactionTemplates[1].Index)
-		assert.Equal(t, 2, resp.Project.TransactionTemplates[2].Index)
-	})
-
-	t.Run("Get transaction templates for non-existent project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp GetProjectTransactionTemplatesResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			QueryGetProjectTransactionTemplates,
-			&resp,
-			client.Var("projectId", badID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Delete transaction template without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		template := createTransactionTemplate(t, c, project)
-
-		var resp DeleteTransactionTemplateResponse
-
-		err := c.Post(
-			MutationDeleteTransactionTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", template.ID),
-		)
-
-		assert.Error(t, err)
-		assert.Empty(t, resp.DeleteTransactionTemplate)
-	})
-
-	t.Run("Delete transaction template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		template := createTransactionTemplate(t, c, project)
-
-		var resp DeleteTransactionTemplateResponse
-
-		err := c.Post(
-			MutationDeleteTransactionTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", template.ID),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, template.ID, resp.DeleteTransactionTemplate)
-	})
-}
-
-func TestTransactionExecutions(t *testing.T) {
-	t.Parallel()
-
-	t.Run("Create execution for non-existent project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		badID := uuid.New().String()
-
-		var resp CreateTransactionExecutionResponse
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", badID),
-			client.Var("script", "transaction { execute { log(\"Hello, World!\") } }"),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Create execution without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = "transaction { execute { log(\"Hello, World!\") } }"
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Create execution", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = "transaction { execute { log(\"Hello, World!\") } }"
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Empty(t, resp.CreateTransactionExecution.Errors)
-		assert.Contains(t, resp.CreateTransactionExecution.Logs, "\"Hello, World!\"")
-		assert.Equal(t, script, resp.CreateTransactionExecution.Script)
-	})
-
-	t.Run("Multiple executions", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateTransactionExecutionResponse
-
-		const script = "transaction { prepare(signer: AuthAccount) { AuthAccount(payer: signer) } }"
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.Var("signers", []string{project.Accounts[0].Address}),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Empty(t, respA.CreateTransactionExecution.Errors)
-		require.Len(t, respA.CreateTransactionExecution.Events, 6)
-
-		eventA := respA.CreateTransactionExecution.Events[5]
-
-		// first account should have address 0x0a
-		assert.Equal(t, "flow.AccountCreated", eventA.Type)
-		assert.JSONEq(t,
-			`{"type":"Address","value":"0x000000000000000a"}`,
-			eventA.Values[0],
-		)
-
-		var respB CreateTransactionExecutionResponse
-
-		err = c.Post(
-			MutationCreateTransactionExecution,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.Var("signers", []string{project.Accounts[0].Address}),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Empty(t, respB.CreateTransactionExecution.Errors)
-		require.Len(t, respB.CreateTransactionExecution.Events, 6)
-
-		eventB := respB.CreateTransactionExecution.Events[5]
-
-		// second account should have address 0x07
-		assert.Equal(t, "flow.AccountCreated", eventB.Type)
-		assert.JSONEq(t,
-			`{"type":"Address","value":"0x000000000000000b"}`,
-			eventB.Values[0],
-		)
-	})
-
-	t.Run("Multiple executions with reset", func(t *testing.T) {
-		t.Parallel()
-		// manually construct resolver
-		c := newClient()
-		project := createProject(t, c)
-
-		var respA CreateTransactionExecutionResponse
-
-		const script = "transaction { prepare(signer: AuthAccount) { AuthAccount(payer: signer) } }"
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.Var("signers", []string{project.Accounts[0].Address}),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Empty(t, respA.CreateTransactionExecution.Errors)
-		require.Len(t, respA.CreateTransactionExecution.Events, 6)
-
-		eventA := respA.CreateTransactionExecution.Events[5]
-
-		// first account should have address 0x0a
-		assert.Equal(t, "flow.AccountCreated", eventA.Type)
-		assert.JSONEq(t,
-			`{"type":"Address","value":"0x000000000000000a"}`,
-			eventA.Values[0],
-		)
-
-		_, err = c.projects.Reset(&model.Project{
-			ID: uuid.MustParse(project.ID),
-		})
-		require.NoError(t, err)
-
-		var respB CreateTransactionExecutionResponse
-
-		err = c.Post(
-			MutationCreateTransactionExecution,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.Var("signers", []string{project.Accounts[0].Address}),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Len(t, respB.CreateTransactionExecution.Events, 6)
-
-		eventB := respB.CreateTransactionExecution.Events[5]
-
-		// second account should have address 0x0a again due to reset
-		assert.Equal(t, "flow.AccountCreated", eventB.Type)
-		assert.JSONEq(t,
-			`{"type":"Address","value":"0x000000000000000a"}`,
-			eventB.Values[0],
-		)
-	})
-
-	t.Run("invalid (parse error)", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = `
-          transaction(a: Int) {
-        `
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "unexpected token: EOF",
-					StartPosition: &model.ProgramPosition{
-						Offset: 41,
-						Line:   3,
-						Column: 8,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 41,
-						Line:   3,
-						Column: 8,
-					},
-				},
-			},
-			resp.CreateTransactionExecution.Errors,
-		)
-		require.Empty(t, resp.CreateTransactionExecution.Logs)
-	})
-
-	t.Run("invalid (semantic error)", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = `
-          transaction { execute { XYZ } }
-        `
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "cannot find variable in this scope: `XYZ`",
-					StartPosition: &model.ProgramPosition{
-						Offset: 35,
-						Line:   2,
-						Column: 34,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 37,
-						Line:   2,
-						Column: 36,
-					},
-				},
-			},
-			resp.CreateTransactionExecution.Errors,
-		)
-		require.Empty(t, resp.CreateTransactionExecution.Logs)
-	})
-
-	t.Run("invalid (run-time error)", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = `
-          transaction { execute { panic("oh no") } }
-        `
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "panic: oh no",
-					StartPosition: &model.ProgramPosition{
-						Offset: 35,
-						Line:   2,
-						Column: 34,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 48,
-						Line:   2,
-						Column: 47,
-					},
-				},
-			},
-			resp.CreateTransactionExecution.Errors,
-		)
-		require.Empty(t, resp.CreateTransactionExecution.Logs)
-	})
-
-	t.Run("exceeding computation limit", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = `
-          transaction {
-              execute {
-                  var i = 0
-                  while i < 1_000_000 {
-                      i = i + 1
-                  }
-              }
-          }
-        `
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, script, resp.CreateTransactionExecution.Script)
-		require.Equal(t,
-			"[Error Code: 1110] computation exceeds limit (100000)",
-			resp.CreateTransactionExecution.Errors[0].Message,
-		)
-	})
-
-	t.Run("argument", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateTransactionExecutionResponse
-
-		const script = `
-          transaction(a: Int) {
-              execute {
-                  log(a)
-              }
-          }
-        `
-
-		err := c.Post(
-			MutationCreateTransactionExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.Var("arguments", []string{
-				`{"type": "Int", "value": "42"}`,
-			}),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		require.Empty(t, resp.CreateTransactionExecution.Errors)
-		require.Equal(t, resp.CreateTransactionExecution.Logs, []string{"42"})
-	})
-}
-
-func TestScriptTemplates(t *testing.T) {
-	t.Parallel()
-	t.Run("Create script template without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptTemplateResponse
-
-		err := c.Post(
-			MutationCreateScriptTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "bar"),
-		)
-
-		assert.Error(t, err)
-		assert.Empty(t, resp.CreateScriptTemplate.ID)
-	})
-
-	t.Run("Create script template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptTemplateResponse
-
-		err := c.Post(
-			MutationCreateScriptTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "bar"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.NotEmpty(t, resp.CreateScriptTemplate.ID)
-		assert.Equal(t, "foo", resp.CreateScriptTemplate.Title)
-		assert.Equal(t, "bar", resp.CreateScriptTemplate.Script)
-	})
-
-	t.Run("Get script template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateScriptTemplateResponse
-
-		err := c.Post(
-			MutationCreateScriptTemplate,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "bar"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		var respB GetScriptTemplateResponse
-
-		err = c.Post(
-			QueryGetScriptTemplate,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", respA.CreateScriptTemplate.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, respA.CreateScriptTemplate.ID, respB.ScriptTemplate.ID)
-		assert.Equal(t, respA.CreateScriptTemplate.Script, respB.ScriptTemplate.Script)
-	})
-
-	t.Run("Get non-existent script template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp GetScriptTemplateResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			QueryGetScriptTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", badID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Update script template without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateScriptTemplateResponse
-
-		err := c.Post(
-			MutationCreateScriptTemplate,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "apple"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		templateID := respA.CreateScriptTemplate.ID
-
-		var respB UpdateScriptTemplateResponse
-
-		err = c.Post(
-			MutationUpdateScriptTemplateScript,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.Var("script", "orange"),
-		)
-		assert.Error(t, err)
-	})
-
-	t.Run("Update script template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA CreateScriptTemplateResponse
-
-		err := c.Post(
-			MutationCreateScriptTemplate,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", "foo"),
-			client.Var("script", "apple"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		templateID := respA.CreateScriptTemplate.ID
-
-		var respB UpdateScriptTemplateResponse
-
-		err = c.Post(
-			MutationUpdateScriptTemplateScript,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.Var("script", "orange"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, respA.CreateScriptTemplate.ID, respB.UpdateScriptTemplate.ID)
-		assert.Equal(t, respA.CreateScriptTemplate.Index, respB.UpdateScriptTemplate.Index)
-		assert.Equal(t, "orange", respB.UpdateScriptTemplate.Script)
-
-		var respC UpdateScriptTemplateResponse
-
-		err = c.Post(
-			MutationUpdateScriptTemplateIndex,
-			&respC,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.Var("index", 1),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, respA.CreateScriptTemplate.ID, respC.UpdateScriptTemplate.ID)
-		assert.Equal(t, 1, respC.UpdateScriptTemplate.Index)
-		assert.Equal(t, respB.UpdateScriptTemplate.Script, respC.UpdateScriptTemplate.Script)
-	})
-
-	t.Run("Update non-existent script template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp UpdateScriptTemplateResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			MutationUpdateScriptTemplateScript,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", badID),
-			client.Var("script", "bar"),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Get script templates for project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		templateIDA := createScriptTemplate(t, c, project)
-		templateIDB := createScriptTemplate(t, c, project)
-		templateIDC := createScriptTemplate(t, c, project)
-
-		var resp GetProjectScriptTemplatesResponse
-
-		err := c.Post(
-			QueryGetProjectScriptTemplates,
-			&resp,
-			client.Var("projectId", project.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Len(t, resp.Project.ScriptTemplates, 3)
-		assert.Equal(t, templateIDA, resp.Project.ScriptTemplates[0].ID)
-		assert.Equal(t, templateIDB, resp.Project.ScriptTemplates[1].ID)
-		assert.Equal(t, templateIDC, resp.Project.ScriptTemplates[2].ID)
-
-		assert.Equal(t, 0, resp.Project.ScriptTemplates[0].Index)
-		assert.Equal(t, 1, resp.Project.ScriptTemplates[1].Index)
-		assert.Equal(t, 2, resp.Project.ScriptTemplates[2].Index)
-	})
-
-	t.Run("Get script templates for non-existent project", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var resp GetProjectScriptTemplatesResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-
-			QueryGetProjectScriptTemplates,
-			&resp,
-			client.Var("projectId", badID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Delete script template without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		templateID := createScriptTemplate(t, c, project)
-
-		var resp DeleteScriptTemplateResponse
-
-		err := c.Post(
-			MutationDeleteScriptTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Delete script template", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		templateID := createScriptTemplate(t, c, project)
-
-		var resp DeleteScriptTemplateResponse
-
-		err := c.Post(
-			MutationDeleteScriptTemplate,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("templateId", templateID),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, templateID, resp.DeleteScriptTemplate)
-	})
-}
-
-func TestAccounts(t *testing.T) {
-	t.Parallel()
-	t.Run("Get account", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-		account := project.Accounts[0]
-
-		var resp GetAccountResponse
-
-		err := c.Post(
-			QueryGetAccount,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, account.ID, resp.Account.ID)
-	})
-
-	t.Run("Get non-existent account", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp GetAccountResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			QueryGetAccount,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", badID),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Update account draft code without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-		account := project.Accounts[0]
-
-		var respA GetAccountResponse
-
-		err := c.Post(
-			QueryGetAccount,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, "", respA.Account.DraftCode)
-
-		var respB UpdateAccountResponse
-
-		err = c.Post(
-			MutationUpdateAccountDraftCode,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", "bar"),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Update account draft code", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-		account := project.Accounts[0]
-
-		var respA GetAccountResponse
-
-		err := c.Post(
-			QueryGetAccount,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, "", respA.Account.DraftCode)
-
-		var respB UpdateAccountResponse
-
-		err = c.Post(
-			MutationUpdateAccountDraftCode,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", "bar"),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, "bar", respB.UpdateAccount.DraftCode)
-	})
-
-	t.Run("Update account invalid deployed code", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-		account := project.Accounts[0]
-
-		var respA GetAccountResponse
-
-		err := c.Post(
-			QueryGetAccount,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, "", respA.Account.DeployedCode)
-
-		var respB UpdateAccountResponse
-
-		err = c.Post(
-			MutationUpdateAccountDeployedCode,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", "INVALID CADENCE"),
-		)
-
-		assert.Error(t, err)
-		assert.Equal(t, "", respB.UpdateAccount.DeployedCode)
-	})
-
-	t.Run("Update account deployed code without permission", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		account := project.Accounts[0]
-
-		var resp UpdateAccountResponse
-
-		const contract = "pub contract Foo {}"
-
-		err := c.Post(
-			MutationUpdateAccountDeployedCode,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", contract),
-		)
-
-		assert.Error(t, err)
-	})
-
-	t.Run("Update account deployed code", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		account := project.Accounts[0]
-
-		var respA GetAccountResponse
-
-		err := c.Post(
-			QueryGetAccount,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, "", respA.Account.DeployedCode)
-
-		var respB UpdateAccountResponse
-
-		const contract = "pub contract Foo {}"
-
-		err = c.Post(
-			MutationUpdateAccountDeployedCode,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", contract),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, contract, respB.UpdateAccount.DeployedCode)
-		assert.Contains(t, respB.UpdateAccount.DeployedContracts, "Foo")
-	})
-
-	t.Run("Update account and redeploy code", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		account := project.Accounts[0]
-
-		var respA GetAccountResponse
-
-		err := c.Post(
-			QueryGetAccount,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, "", respA.Account.DeployedCode)
-
-		var respB UpdateAccountResponse
-
-		const contract = "pub contract Foo {}"
-
-		err = c.Post(
-			MutationUpdateAccountDeployedCode,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", contract),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, contract, respB.UpdateAccount.DeployedCode)
-		assert.Contains(t, respB.UpdateAccount.DeployedContracts, "Foo")
-
-		var respC UpdateAccountResponse
-
-		const contract2 = "pub contract Bar {}"
-
-		err = c.Post(
-			MutationUpdateAccountDeployedCode,
-			&respC,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", account.ID),
-			client.Var("code", contract2),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, contract2, respC.UpdateAccount.DeployedCode)
-		assert.Contains(t, respC.UpdateAccount.DeployedContracts, "Bar")
-
-		var resp GetProjectResponse
-
-		err = c.Post(
-			QueryGetProject,
-			&resp,
-			client.Var("projectId", project.ID),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, project.ID, resp.Project.ID)
-		assert.Len(t, resp.Project.Accounts, 5)
-
-		assert.Equal(t, contract2, resp.Project.Accounts[0].DeployedCode)
-	})
-
-	t.Run("Update non-existent account", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp UpdateAccountResponse
-
-		badID := uuid.New().String()
-
-		err := c.Post(
-			MutationUpdateAccountDraftCode,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("accountId", badID),
-			client.Var("script", "bar"),
-		)
-
-		assert.Error(t, err)
-	})
-}
-
-const counterContract = `
-  pub contract Counting {
-
-      pub event CountIncremented(count: Int)
-
-      pub resource Counter {
-          pub var count: Int
-
-          init() {
-              self.count = 0
-          }
-
-          pub fun add(_ count: Int) {
-              self.count = self.count + count
-              emit CountIncremented(count: self.count)
-          }
-      }
-
-      pub fun createCounter(): @Counter {
-          return <-create Counter()
-      }
-  }
-`
-
-// generateAddTwoToCounterScript generates a script that increments a counter.
-// If no counter exists, it is created.
-func generateAddTwoToCounterScript(counterAddress string) string {
-	return fmt.Sprintf(
-		`
-            import 0x%s
-
-            transaction {
-
-                prepare(signer: AuthAccount) {
-                    if signer.borrow<&Counting.Counter>(from: /storage/counter) == nil {
-                        signer.save(<-Counting.createCounter(), to: /storage/counter)
-                    }
-
-                    signer.borrow<&Counting.Counter>(from: /storage/counter)!.add(2)
-                }
-            }
-        `,
-		counterAddress,
-	)
-}
-
-func TestContractInteraction(t *testing.T) {
-	t.Parallel()
-	c := newClient()
-
-	project := createProject(t, c)
-
-	accountA := project.Accounts[0]
-	accountB := project.Accounts[1]
-
-	var respA UpdateAccountResponse
-
-	err := c.Post(
-		MutationUpdateAccountDeployedCode,
-		&respA,
-		client.Var("projectId", project.ID),
-		client.Var("accountId", accountA.ID),
-		client.Var("code", counterContract),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-
-	assert.Equal(t, counterContract, respA.UpdateAccount.DeployedCode)
-
-	addScript := generateAddTwoToCounterScript(accountA.Address)
-
-	var respB CreateTransactionExecutionResponse
-
-	err = c.Post(
-		MutationCreateTransactionExecution,
-		&respB,
-		client.Var("projectId", project.ID),
-		client.Var("script", addScript),
-		client.Var("signers", []string{accountB.Address}),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-
-	assert.Empty(t, respB.CreateTransactionExecution.Errors)
-}
-
-func TestContractImport(t *testing.T) {
-	t.Parallel()
-	c := newClient()
-
-	project := createProject(t, c)
-
-	accountA := project.Accounts[0]
-	accountB := project.Accounts[1]
-
-	contractA := `
-	pub contract HelloWorldA {
-		pub var A: String
-		pub init() { self.A = "HelloWorldA" }
-	}`
-
-	contractB := `
-	import HelloWorldA from 0x01
-	pub contract HelloWorldB {
-		pub init() {
-			log(HelloWorldA.A)
-		}
-	}`
-
-	var respA UpdateAccountResponse
-
-	err := c.Post(
-		MutationUpdateAccountDeployedCode,
-		&respA,
-		client.Var("projectId", project.ID),
-		client.Var("accountId", accountA.ID),
-		client.Var("code", contractA),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-	assert.Equal(t, contractA, respA.UpdateAccount.DeployedCode)
-
-	var respB UpdateAccountResponse
-
-	err = c.Post(
-		MutationUpdateAccountDeployedCode,
-		&respB,
-		client.Var("projectId", project.ID),
-		client.Var("accountId", accountB.ID),
-		client.Var("code", contractB),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-}
-
-func TestAccountStorage(t *testing.T) {
-	t.Parallel()
-	c := newClient()
-
-	project := createProject(t, c)
-	account := project.Accounts[0]
-
-	var accResp GetAccountResponse
-
-	err := c.Post(
-		QueryGetAccount,
-		&accResp,
-		client.Var("projectId", project.ID),
-		client.Var("accountId", account.ID),
-	)
-	require.NoError(t, err)
-
-	assert.Equal(t, account.ID, accResp.Account.ID)
-	assert.Equal(t, `{}`, accResp.Account.State)
-
-	var resp CreateTransactionExecutionResponse
-
-	const script = `
-		transaction {
-		  prepare(signer: AuthAccount) {
-			  	signer.save("storage value", to: /storage/storageTest)
- 				signer.link<&String>(/public/publicTest, target: /storage/storageTest)
-				signer.link<&String>(/private/privateTest, target: /storage/storageTest)
-		  }
-   		}`
-
-	err = c.Post(
-		MutationCreateTransactionExecution,
-		&resp,
-		client.Var("projectId", project.ID),
-		client.Var("script", script),
-		client.Var("signers", []string{account.Address}),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-
-	err = c.Post(
-		QueryGetAccount,
-		&accResp,
-		client.Var("projectId", project.ID),
-		client.Var("accountId", account.ID),
-	)
-	require.NoError(t, err)
-
-	assert.Equal(t, account.ID, accResp.Account.ID)
-	assert.NotEmpty(t, accResp.Account.State)
-
-	type accountStorage struct {
-		Private map[string]any
-		Public  map[string]any
-		Storage map[string]any
-	}
-
-	var accStorage accountStorage
-	err = json.Unmarshal([]byte(accResp.Account.State), &accStorage)
-	require.NoError(t, err)
-
-	assert.Equal(t, "storage value", accStorage.Storage["storageTest"])
-	assert.NotEmpty(t, accStorage.Private["privateTest"])
-	assert.NotEmpty(t, accStorage.Public["publicTest"])
-
-	assert.NotContains(t, accStorage.Public, "flowTokenBalance")
-	assert.NotContains(t, accStorage.Public, "flowTokenReceiver")
-	assert.NotContains(t, accStorage.Storage, "flowTokenVault")
-}
-
-func TestAuthentication(t *testing.T) {
-	t.Parallel()
-	t.Run("Migrate legacy auth", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var respA UpdateProjectResponse
-
-		oldSessionCookie := c.SessionCookie()
-
-		// clear session cookie before making request
-		c.ClearSessionCookie()
-
-		err := c.Post(
-			MutationUpdateProjectPersist,
-			&respA,
-			client.Var("projectId", project.ID),
-			client.Var("title", project.Title),
-			client.Var("description", project.Description),
-			client.Var("readme", project.Readme),
-			client.Var("persist", true),
-			client.AddCookie(legacyauth.MockProjectSessionCookie(project.ID, project.Secret)),
-		)
-		require.NoError(t, err)
-
-		assert.Equal(t, project.ID, respA.UpdateProject.ID)
-		assert.Equal(t, project.Title, respA.UpdateProject.Title)
-		assert.Equal(t, project.Description, respA.UpdateProject.Description)
-		assert.Equal(t, project.Readme, respA.UpdateProject.Readme)
-		assert.True(t, respA.UpdateProject.Persist)
-
-		// a new session cookie should be set
-		require.NotNil(t, c.SessionCookie())
-		assert.NotEqual(t, oldSessionCookie.Value, c.SessionCookie().Value)
-
-		var respB UpdateProjectResponse
-
-		err = c.Post(
-			MutationUpdateProjectPersist,
-			&respB,
-			client.Var("projectId", project.ID),
-			client.Var("title", project.Title),
-			client.Var("description", project.Description),
-			client.Var("readme", project.Readme),
-			client.Var("persist", false),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		// should be able to perform update using new session cookie
-		assert.Equal(t, project.ID, respB.UpdateProject.ID)
-		assert.Equal(t, project.Title, respB.UpdateProject.Title)
-		assert.Equal(t, project.Description, respB.UpdateProject.Description)
-		assert.Equal(t, project.Readme, respB.UpdateProject.Readme)
-		assert.False(t, respB.UpdateProject.Persist)
-	})
-
-	t.Run("Create project with malformed session cookie", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		var respA CreateProjectResponse
-
-		malformedCookie := http.Cookie{
-			Name:  sessionName,
-			Value: "foo",
-		}
-
-		err := c.Post(
-			MutationCreateProject,
-			&respA,
-			client.Var("title", "foo"),
-			client.Var("description", "desc"),
-			client.Var("readme", "rtfm"),
-			client.Var("seed", 42),
-			client.AddCookie(&malformedCookie),
-		)
-		require.NoError(t, err)
-
-		projectID := respA.CreateProject.ID
-		projectTitle := respA.CreateProject.Title
-		projectDescription := respA.CreateProject.Description
-		projectReadme := respA.CreateProject.Readme
-
-		assert.NotEmpty(t, projectID)
-		assert.Equal(t, 42, respA.CreateProject.Seed)
-
-		// session cookie should be overwritten with new value
-		assert.NotNil(t, c.SessionCookie())
-
-		var respB UpdateProjectResponse
-
-		err = c.Post(
-			MutationUpdateProjectPersist,
-			&respB,
-			client.Var("projectId", projectID),
-			client.Var("title", projectTitle),
-			client.Var("description", projectDescription),
-			client.Var("readme", projectReadme),
-			client.Var("persist", true),
-			client.AddCookie(c.SessionCookie()),
-		)
-		require.NoError(t, err)
-
-		// should be able to perform update using new session cookie
-		assert.Equal(t, projectID, respB.UpdateProject.ID)
-		assert.Equal(t, projectTitle, respB.UpdateProject.Title)
-		assert.Equal(t, projectDescription, respB.UpdateProject.Description)
-		assert.Equal(t, projectReadme, respB.UpdateProject.Readme)
-		assert.True(t, respB.UpdateProject.Persist)
-	})
-
-	t.Run("Update project with malformed session cookie", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp UpdateProjectResponse
-
-		malformedCookie := http.Cookie{
-			Name:  sessionName,
-			Value: "foo",
-		}
-
-		c.ClearSessionCookie()
-
-		err := c.Post(
-			MutationUpdateProjectPersist,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("persist", true),
-			client.AddCookie(&malformedCookie),
-		)
-
-		assert.Error(t, err)
-
-		// session cookie should not be set
-		assert.Nil(t, c.SessionCookie())
-	})
-
-	t.Run("Update project with invalid session cookie", func(t *testing.T) {
-		t.Parallel()
-		c := newClient()
-
-		projectA := createProject(t, c)
-		_ = createProject(t, c)
-
-		cookieB := c.SessionCookie()
-
-		var resp UpdateProjectResponse
-
-		err := c.Post(
-			MutationUpdateProjectPersist,
-			&resp,
-			client.Var("projectId", projectA.ID),
-			client.Var("persist", true),
-			client.AddCookie(cookieB),
-		)
-
-		// should not be able to update project A with cookie B
-		assert.Error(t, err)
-	})
-}
-
-func TestScriptExecutions(t *testing.T) {
-	t.Parallel()
-
-	t.Run("valid, no return value", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = "pub fun main() { }"
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		require.Empty(t, resp.CreateScriptExecution.Errors)
-	})
-
-	t.Run("invalid (parse error)", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = "pub fun main() {"
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		assert.Equal(t, script, resp.CreateScriptExecution.Script)
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "expected token '}'",
-					StartPosition: &model.ProgramPosition{
-						Offset: 16,
-						Line:   1,
-						Column: 16,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 16,
-						Line:   1,
-						Column: 16,
-					},
-				},
-			},
-			resp.CreateScriptExecution.Errors,
-		)
-	})
-
-	t.Run("invalid (semantic error)", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = "pub fun main() { XYZ }"
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		assert.Equal(t, script, resp.CreateScriptExecution.Script)
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "cannot find variable in this scope: `XYZ`",
-					StartPosition: &model.ProgramPosition{
-						Offset: 17,
-						Line:   1,
-						Column: 17,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 19,
-						Line:   1,
-						Column: 19,
-					},
-				},
-			},
-			resp.CreateScriptExecution.Errors,
-		)
-	})
-
-	t.Run("invalid (run-time error)", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = "pub fun main() { panic(\"oh no\") }"
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		assert.Equal(t, script, resp.CreateScriptExecution.Script)
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "panic: oh no",
-					StartPosition: &model.ProgramPosition{
-						Offset: 17,
-						Line:   1,
-						Column: 17,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 30,
-						Line:   1,
-						Column: 30,
-					},
-				},
-			},
-			resp.CreateScriptExecution.Errors,
-		)
-	})
-
-	t.Run("exceeding computation limit", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = `
-          pub fun main() {
-              var i = 0
-              while i < 1_000_000 {
-                  i = i + 1
-              }
-          }
-        `
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		assert.Equal(t, script, resp.CreateScriptExecution.Script)
-		require.Equal(t,
-			[]model.ProgramError{
-				{
-					Message: "[Error Code: 1110] computation exceeds limit (100000)",
-					StartPosition: &model.ProgramPosition{
-						Offset: 106,
-						Line:   5,
-						Column: 18,
-					},
-					EndPosition: &model.ProgramPosition{
-						Offset: 114,
-						Line:   5,
-						Column: 26,
-					},
-				},
-			},
-			resp.CreateScriptExecution.Errors,
-		)
-	})
-
-	t.Run("return address", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = "pub fun main(): Address { return 0x1 as Address }"
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		assert.Equal(t, script, resp.CreateScriptExecution.Script)
-		require.Empty(t, resp.CreateScriptExecution.Errors)
-		assert.JSONEq(t,
-			`{"type":"Address","value":"0x0000000000000001"}`,
-			resp.CreateScriptExecution.Value,
-		)
-	})
-
-	t.Run("argument", func(t *testing.T) {
-		t.Parallel()
-
-		c := newClient()
-
-		project := createProject(t, c)
-
-		var resp CreateScriptExecutionResponse
-
-		const script = "pub fun main(a: Int): Int { return a + 1 }"
-
-		err := c.Post(
-			MutationCreateScriptExecution,
-			&resp,
-			client.Var("projectId", project.ID),
-			client.Var("script", script),
-			client.Var("arguments", []string{
-				`{"type":"Int","value":"2"}`,
-			}),
-			client.AddCookie(c.SessionCookie()),
-		)
-
-		require.NoError(t, err)
-		assert.Equal(t, script, resp.CreateScriptExecution.Script)
-		require.Empty(t, resp.CreateScriptExecution.Errors)
-		assert.JSONEq(t,
-			`{"type":"Int","value":"3"}`,
-			resp.CreateScriptExecution.Value,
-		)
-	})
-}
-*/
-
-type Client struct {
-	client        *client.Client
-	resolver      *playground.Resolver
-	sessionCookie *http.Cookie
-	projects      *blockchain.Projects
-	store         storage.Store
-}
-
-func (c *Client) Post(query string, response interface{}, options ...client.Option) error {
-	w := httptest.NewRecorder()
-
-	err := c.client.Post(w, query, response, options...)
-
-	for _, cookie := range w.Result().Cookies() {
-		if cookie.Name == sessionName {
-			c.sessionCookie = cookie
-		}
-	}
-
-	return err
-}
-
-func (c *Client) MustPost(query string, response interface{}, options ...client.Option) {
-	err := c.Post(query, response, options...)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (c *Client) SessionCookie() *http.Cookie {
-	return c.sessionCookie
-}
-
-func (c *Client) ClearSessionCookie() {
-	c.sessionCookie = nil
-}
-
-const sessionName = "flow-playground-test"
-
-var version, _ = semver.NewVersion("0.1.0")
-
-// keep same instance of store due to connection pool exhaustion
-var store storage.Store
-
-func newStore() storage.Store {
-	if store != nil {
-		return store
-	}
-
-	if strings.EqualFold(os.Getenv("FLOW_STORAGEBACKEND"), storage.PostgreSQL) {
-		var datastoreConf storage.DatabaseConfig
-		if err := envconfig.Process("FLOW_DB", &datastoreConf); err != nil {
-			panic(err)
-		}
-
-		store = storage.NewPostgreSQL(&datastoreConf)
-	} else {
-		store = storage.NewSqlite()
-	}
-
-	return store
-}
-
-func newClient() *Client {
-	store := newStore()
-	authenticator := auth.NewAuthenticator(store, sessionName)
-	chain := blockchain.NewProjects(store, initAccounts)
-	resolver := playground.NewResolver(version, store, authenticator, chain)
-
-	c := newClientWithResolver(resolver)
-	c.store = store
-	c.projects = chain
-	return c
-}
-
-func newClientWithResolver(resolver *playground.Resolver) *Client {
-	router := chi.NewRouter()
-	router.Use(httpcontext.Middleware())
-	router.Use(legacyauth.MockProjectSessions())
-
-	localHub := sentry.CurrentHub().Clone()
-	logger := logrus.StandardLogger()
-	entry := logrus.NewEntry(logger)
-	router.Handle("/", playground.GraphQLHandler(resolver, errors.Middleware(entry, localHub)))
-
-	return &Client{
-		client:   client.New(router),
-		resolver: resolver,
-	}
-}
-
-func createProject(t *testing.T, c *Client) Project {
-	var resp CreateProjectResponse
-
-	err := c.Post(
-		MutationCreateProject,
-		&resp,
-		client.Var("title", "foo"),
-		client.Var("seed", 42),
-		client.Var("description", "desc"),
-		client.Var("readme", "rtfm"),
-		client.Var("numberOfAccounts", 5),
-		client.Var("transactionTemplates", []string{}),
-	)
-	require.NoError(t, err)
-
-	proj := resp.CreateProject
-	internalProj := c.resolver.LastCreatedProject()
-
-	proj.Secret = internalProj.Secret.String()
-
-	return proj
-}
-
-func createTransactionTemplate(t *testing.T, c *Client, project Project) TransactionTemplate {
-	var resp CreateTransactionTemplateResponse
-
-	err := c.Post(
-		MutationCreateTransactionTemplate,
-		&resp,
-		client.Var("projectId", project.ID),
-		client.Var("title", "foo"),
-		client.Var("script", "bar"),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-
-	return resp.CreateTransactionTemplate
-}
-
-func createScriptTemplate(t *testing.T, c *Client, project Project) string {
-	var resp CreateScriptTemplateResponse
-
-	err := c.Post(
-		MutationCreateScriptTemplate,
-		&resp,
-		client.Var("projectId", project.ID),
-		client.Var("title", "foo"),
-		client.Var("script", "bar"),
-		client.AddCookie(c.SessionCookie()),
-	)
-	require.NoError(t, err)
-
-	return resp.CreateScriptTemplate.ID
-}
-
-// todo add tests for:
-// - failed transactions with successful transactions work (bootstrap works)??
-// - assert we don't leak any internal model data to API
