+++ test/client/client_test.go
@@ -20,14 +20,13 @@ package client_test
 
 import (
 	"encoding/json"
+	client2 "github.com/dapperlabs/flow-playground-api/test/client"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-
-	"github.com/dapperlabs/flow-playground-api/client"
 )
 
 func TestClient(t *testing.T) {
@@ -48,7 +47,7 @@ func TestClient(t *testing.T) {
 		}
 	})
 
-	c := client.New(h)
+	c := client2.New(h)
 
 	var resp struct {
 		Name string
@@ -56,7 +55,7 @@ func TestClient(t *testing.T) {
 
 	w := httptest.NewRecorder()
 
-	err := c.Post(w, "user(id:$id){name}", &resp, client.Var("id", 1))
+	err := c.Post(w, "user(id:$id){name}", &resp, client2.Var("id", 1))
 	require.NoError(t, err)
 
 	require.Equal(t, "bob", resp.Name)
@@ -69,14 +68,14 @@ func TestAddHeader(t *testing.T) {
 		_, _ = w.Write([]byte(`{}`))
 	})
 
-	c := client.New(h)
+	c := client2.New(h)
 
 	var resp struct{}
 
 	w := httptest.NewRecorder()
 
 	err := c.Post(w, "{ id }", &resp,
-		client.AddHeader("Test-Key", "ASDF"),
+		client2.AddHeader("Test-Key", "ASDF"),
 	)
 	require.NoError(t, err)
 }
@@ -88,7 +87,7 @@ func TestAddClientHeader(t *testing.T) {
 		_, _ = w.Write([]byte(`{}`))
 	})
 
-	c := client.New(h, client.AddHeader("Test-Key", "ASDF"))
+	c := client2.New(h, client2.AddHeader("Test-Key", "ASDF"))
 
 	var resp struct{}
 
@@ -108,14 +107,14 @@ func TestBasicAuth(t *testing.T) {
 		_, _ = w.Write([]byte(`{}`))
 	})
 
-	c := client.New(h)
+	c := client2.New(h)
 
 	var resp struct{}
 
 	w := httptest.NewRecorder()
 
 	err := c.Post(w, "{ id }", &resp,
-		client.BasicAuth("user", "pass"),
+		client2.BasicAuth("user", "pass"),
 	)
 	require.NoError(t, err)
 }
@@ -129,14 +128,14 @@ func TestAddCookie(t *testing.T) {
 		_, _ = w.Write([]byte(`{}`))
 	})
 
-	c := client.New(h)
+	c := client2.New(h)
 
 	var resp struct{}
 
 	w := httptest.NewRecorder()
 
 	err := c.Post(w, "{ id }", &resp,
-		client.AddCookie(&http.Cookie{Name: "foo", Value: "value"}),
+		client2.AddCookie(&http.Cookie{Name: "foo", Value: "value"}),
 	)
 	require.NoError(t, err)
 }
