+++ test/client/errors.go
