+++ test/client/options.go
