+++ test/contract_deployments_test.go
@@ -1,0 +1,3 @@
+package test
+
+// TODO: Test contract deployments similar to transaction executions
